+++ src/components/ProductDetailsModal.js
@@ -1,45 +1,8 @@
 import React from 'react';
 import styled from 'styled-components';
-import {formatDollar, stopPropagation} from '../utils';
+import {formatDollar} from '../utils';
 import {LoadingSpinner} from './LoadingSpinner';
-
-const Modal = styled.div`
-  position: fixed;
-  height: 100%;
-  width: 100%;
-  top: 0;
-  left: 0;
-  background-color: rgba(255,255,255,0.86);
-  z-index: 10;
-`;
-
-const ModalX = styled.div.attrs({
-  className: 'fa fa-times',
-})`
-  position: fixed;
-  top: 0;
-  right: 0;
-  padding: 30px;
-  font-size: 48px;
-  color: #818181;
-  cursor: pointer;
-
-  &:hover {
-    opacity: 0.7;
-  }
-`;
-
-const ModalContents = styled.div`
-  position: relative;
-  top: 20%;
-  min-height: 400px;
-  padding: 100px 60px;
-  width: 850px;
-  margin: 0 auto;
-  background: #FFF;
-  border-radius: 10px;
-  box-shadow: 0 4px 14px 7px rgba(121, 121, 121, 0.10);
-`;
+import {Modal} from './Modal';
 
 const ModalSpinner = styled.div`
   position: relative;
@@ -75,34 +38,29 @@ const ItemDescription = styled.p`
 `;
 
 export const ProductDetailsModal = ({isOpen, close, item}) => {
-  return !isOpen ? null : (
-    <Modal onClick={close}>
-      <ModalContents onClick={stopPropagation}>
-        <ModalX onClick={close} />
-
-        {!item ? (
-          <ModalSpinner>
-            <LoadingSpinner size="large" />
-          </ModalSpinner>
-        ) : (
-          <div>
-            <ItemImage
-              src={item.images.large}
-              alt={item.name}
-            />
-            <ItemName>
-              {item.name}
-            </ItemName>
-            <ItemPrice>
-              {formatDollar(item.price)}
-            </ItemPrice>
-            <ItemDescription>
-              {item.description}
-            </ItemDescription>
-          </div>
-        )}
-
-      </ModalContents>
+  return (
+    <Modal {...{isOpen, close}}>
+      {!item ? (
+        <ModalSpinner>
+          <LoadingSpinner size="large" />
+        </ModalSpinner>
+      ) : (
+        <div>
+          <ItemImage
+            src={item.images.large}
+            alt={item.name}
+          />
+          <ItemName>
+            {item.name}
+          </ItemName>
+          <ItemPrice>
+            {formatDollar(item.price)}
+          </ItemPrice>
+          <ItemDescription>
+            {item.description}
+          </ItemDescription>
+        </div>
+      )}
     </Modal>
-  )
+  );
 };
